fix: resolve merge artifacts from iteration 2 lock_or_recover + i64 widening
- Remove duplicate sync_utils module declaration and lock_or_recover imports
- Align playtime type to i64 in pad_gaugelog_with_zeros and scoring
- Remove destructive lr.init() in receive_updated_play_config that
  overwrote user's hispeed/scroll config

+++ crates/rubato-play/src/bms_player/scoring.rs
@@ -356,7 +356,7 @@ impl BMSPlayer {
         self.judge.past_notes()
     }
 
-    pub fn playtime(&self) -> i32 {
+    pub fn playtime(&self) -> i64 {
         self.playtime
     }
